uvm: "state" subcommand and "-s" option for "ssh" subcommand
The new "state" subcommand returns a short string describing the current state of the VM ("running", "paused", or "powered off").

The new "-s"/"--start" switch for the "ssh" subcommand tells it to automatically start or resume the VM as necessary and wait for it to be available over SSH. This makes it easier to quickly get into the VM no matter what state it's in.

Finally, a few tweaks have been made to the state management subcommands ("start", "stop", "pause", and "resume"). These make it so that they're aware of the current state of the VM and only perform any actions when they can. For example, the "start" subcommand will do nothing and succeed if the VM is already running. It will also do nothing but fail when the VM is paused. Otherwise, the VM is powered off, so it will start it.